refactor(mocker): regroup replay runtime
