+++ lib/mocker/src/replay/collector.rs
@@ -2,17 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use std::collections::HashMap;
-use std::fs::File;
-use std::io::{BufRead, BufReader};
-use std::path::Path;
-use std::time::Instant;
 
-use anyhow::{Context, Result, anyhow, bail};
+use serde::Serialize;
 use serde::ser::{SerializeMap, Serializer};
-use serde::{Deserialize, Serialize};
 use uuid::Uuid;
-
-use crate::common::protocols::{DirectRequest, EngineType, MockEngineArgs, WorkerType};
 
 #[derive(Debug, Clone)]
 pub struct TraceSimulationReport {
@@ -369,197 +362,6 @@ impl TraceCollector {
     }
 }
 
-#[derive(Debug, Deserialize)]
-struct RawTraceRecord {
-    #[serde(default)]
-    timestamp: Option<f64>,
-    #[serde(default)]
-    created_time: Option<f64>,
-    #[serde(default, alias = "input_tokens")]
-    input_length: Option<usize>,
-    #[serde(default, alias = "output_tokens")]
-    output_length: Option<usize>,
-    #[serde(default)]
-    hash_ids: Option<Vec<u64>>,
-}
-
-pub fn simulate_trace_file(
-    args: MockEngineArgs,
-    trace_path: &Path,
-    num_workers: usize,
-) -> Result<TraceSimulationReport> {
-    validate_offline_replay_args(&args, num_workers)?;
-    let requests = load_trace_requests(trace_path, args.block_size, true)?;
-    let started_at = Instant::now();
-    let report = crate::scheduler::offline::simulate_trace(args, requests, num_workers)?;
-    Ok(report.with_wall_time_ms(started_at.elapsed().as_secs_f64() * 1000.0))
-}
-
-pub fn simulate_concurrency_file(
-    args: MockEngineArgs,
-    trace_path: &Path,
-    max_in_flight: usize,
-    num_workers: usize,
-) -> Result<TraceSimulationReport> {
-    let requests = load_trace_requests(trace_path, args.block_size, false)?;
-    let started_at = Instant::now();
-    let report = simulate_concurrency_requests(args, requests, max_in_flight, num_workers)?;
-    Ok(report.with_wall_time_ms(started_at.elapsed().as_secs_f64() * 1000.0))
-}
-
-pub fn simulate_concurrency_requests(
-    args: MockEngineArgs,
-    requests: Vec<DirectRequest>,
-    max_in_flight: usize,
-    num_workers: usize,
-) -> Result<TraceSimulationReport> {
-    validate_offline_concurrency_args(&args, num_workers, max_in_flight)?;
-    if requests.is_empty() {
-        bail!("concurrency replay requires at least one request");
-    }
-
-    crate::scheduler::offline::simulate_concurrency(args, requests, max_in_flight, num_workers)
-}
-
-fn validate_offline_replay_args(args: &MockEngineArgs, num_workers: usize) -> Result<()> {
-    if num_workers == 0 {
-        bail!("trace replay requires num_workers >= 1");
-    }
-    if args.engine_type != EngineType::Vllm {
-        bail!(
-            "trace replay only supports engine_type=vllm, got {:?}",
-            args.engine_type
-        );
-    }
-    if args.worker_type != WorkerType::Aggregated {
-        bail!(
-            "trace replay only supports aggregated workers, got {:?}",
-            args.worker_type
-        );
-    }
-    if args.dp_size != 1 {
-        bail!(
-            "trace replay only supports data_parallel_size=1, got {}",
-            args.dp_size
-        );
-    }
-
-    Ok(())
-}
-
-fn validate_offline_concurrency_args(
-    args: &MockEngineArgs,
-    num_workers: usize,
-    max_in_flight: usize,
-) -> Result<()> {
-    if max_in_flight == 0 {
-        bail!("concurrency replay requires max_in_flight >= 1");
-    }
-
-    validate_offline_replay_args(args, num_workers)
-}
-
-fn load_trace_requests(
-    trace_path: &Path,
-    trace_block_size: usize,
-    timestamps_required: bool,
-) -> Result<Vec<DirectRequest>> {
-    let file = File::open(trace_path)
-        .with_context(|| format!("failed to open trace file {}", trace_path.display()))?;
-    let reader = BufReader::new(file);
-    let mut requests = Vec::new();
-
-    for (line_idx, line) in reader.lines().enumerate() {
-        let line = line.with_context(|| {
-            format!(
-                "failed to read line {} from {}",
-                line_idx + 1,
-                trace_path.display()
-            )
-        })?;
-        if line.trim().is_empty() {
-            continue;
-        }
-
-        let raw: RawTraceRecord = serde_json::from_str(&line).with_context(|| {
-            format!(
-                "failed to parse line {} from {} as JSON",
-                line_idx + 1,
-                trace_path.display()
-            )
-        })?;
-
-        let input_length = raw
-            .input_length
-            .ok_or_else(|| anyhow!("trace line {} is missing input_length", line_idx + 1))?;
-        let output_length = raw
-            .output_length
-            .ok_or_else(|| anyhow!("trace line {} is missing output_length", line_idx + 1))?;
-        let hash_ids = raw
-            .hash_ids
-            .ok_or_else(|| anyhow!("trace line {} is missing hash_ids", line_idx + 1))?;
-        let arrival_timestamp_ms = if timestamps_required {
-            match raw.timestamp.or(raw.created_time) {
-                Some(timestamp_ms) => Some(timestamp_ms),
-                None => return Err(anyhow!("trace line {} is missing timestamp", line_idx + 1)),
-            }
-        } else {
-            None
-        };
-        let tokens = synthesize_tokens_from_hash_ids(&hash_ids, input_length, trace_block_size)
-            .with_context(|| {
-                format!(
-                    "failed to synthesize tokens from hash_ids on line {}",
-                    line_idx + 1
-                )
-            })?;
-
-        requests.push(DirectRequest {
-            tokens,
-            max_output_tokens: output_length,
-            uuid: Some(Uuid::new_v4()),
-            dp_rank: 0,
-            arrival_timestamp_ms,
-        });
-    }
-
-    if requests.is_empty() {
-        bail!(
-            "trace file {} did not contain any requests",
-            trace_path.display()
-        );
-    }
-
-    Ok(requests)
-}
-
-fn synthesize_tokens_from_hash_ids(
-    hash_ids: &[u64],
-    input_length: usize,
-    trace_block_size: usize,
-) -> Result<Vec<u32>> {
-    let mut tokens = Vec::with_capacity(input_length);
-
-    for &hash_id in hash_ids {
-        let token_id = u32::try_from(hash_id)
-            .map_err(|_| anyhow!("hash_id {hash_id} exceeds u32::MAX for token synthesis"))?;
-        // TODO: Replace this repeated-token expansion with a hash-native prompt representation.
-        tokens.extend((0..trace_block_size).map(|_| token_id));
-        if tokens.len() >= input_length {
-            tokens.truncate(input_length);
-            return Ok(tokens);
-        }
-    }
-
-    bail!(
-        "input_length {} exceeds synthesized capacity {} from {} hash_ids and block_size {}",
-        input_length,
-        hash_ids.len() * trace_block_size,
-        hash_ids.len(),
-        trace_block_size
-    );
-}
-
 fn mean(values: &[f64]) -> f64 {
     if values.is_empty() {
         0.0
@@ -616,42 +418,4 @@ fn std_dev(values: &[f64]) -> f64 {
         .sum::<f64>()
         / values.len() as f64;
     variance.sqrt()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_replay_itl_uses_per_token_gaps() {
-        let mut collector = TraceCollector::default();
-        let uuid = Uuid::from_u128(11);
-
-        collector.on_arrival(uuid, 0.0, 4, 4);
-        collector.on_admit(uuid, 0.0, 0);
-        collector.on_token(uuid, 10.0);
-        collector.on_token(uuid, 11.0);
-        collector.on_token(uuid, 12.0);
-        collector.on_token(uuid, 110.0);
-
-        let report = collector.finish();
-
-        assert!((report.latency.tpot.mean_ms - (100.0 / 3.0)).abs() < 1e-9);
-        assert!((report.latency.itl.distribution.mean_ms - (100.0 / 3.0)).abs() < 1e-9);
-        assert_eq!(report.latency.itl.distribution.median_ms, 1.0);
-        assert_eq!(report.latency.itl.distribution.p75_ms, 98.0);
-        assert_eq!(report.latency.itl.distribution.p90_ms, 98.0);
-        assert_eq!(report.latency.itl.distribution.p95_ms, 98.0);
-        assert_eq!(report.latency.itl.max_ms, 98.0);
-        assert_eq!(report.latency.ttst.min_ms, 1.0);
-        assert_eq!(report.latency.ttst.max_ms, 1.0);
-        assert_eq!(
-            report.latency.output_token_throughput_per_user.min_ms,
-            1000.0 / 98.0
-        );
-        assert_eq!(
-            report.latency.output_token_throughput_per_user.max_ms,
-            1000.0
-        );
-    }
 }
