+++ lib/mocker/src/scheduler/mod.rs
@@ -3,7 +3,6 @@
 
 //! Engine-specific scheduling implementations.
 
-pub(crate) mod offline;
 pub mod sglang;
 pub mod vllm;
